Move bridges to new package
