+++ src/main/java/org/alliancegenome/curation_api/model/entities/Vocabulary.java
@@ -5,7 +5,7 @@ import java.util.List;
 import javax.persistence.*;
 
 import org.alliancegenome.curation_api.base.entity.BaseGeneratedEntity;
-import org.alliancegenome.curation_api.model.entities.ontology.BooleanValueBridge;
+import org.alliancegenome.curation_api.model.bridges.BooleanValueBridge;
 import org.alliancegenome.curation_api.view.View;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.hibernate.envers.Audited;
