+++ src/main/java/org/alliancegenome/curation_api/model/entities/ontology/OntologyTerm.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.*;
 
 import org.alliancegenome.curation_api.base.entity.BaseCurieEntity;
+import org.alliancegenome.curation_api.model.bridges.BooleanValueBridge;
 import org.alliancegenome.curation_api.model.entities.CrossReference;
 import org.alliancegenome.curation_api.view.View;
 import org.hibernate.envers.Audited;
